Moved drivetrains into their own subsystem directory, neatened up ternaries in tank_drive.cpp, removed isDriver boolean.
