+++ core/src/subsystems/drivetrains/tank_drive.cpp
@@ -1,4 +1,4 @@
-#include "../core/include/subsystems/tank_drive.h"
+#include "../core/include/subsystems/drivetrains/tank_drive.h"
 
 TankDrive::TankDrive(motor_group &left_motors, motor_group &right_motors, robot_specs_t &config, OdometryTank *odom)
     : left_motors(left_motors), right_motors(right_motors),
@@ -30,20 +30,12 @@ void TankDrive::stop()
  * 
  * left_motors and right_motors are in "percent": -1.0 -> 1.0
  */
-void TankDrive::drive_tank(double left, double right, int power, bool isdriver)
+void TankDrive::drive_tank(double left, double right, int power)
 {
   left = modify_inputs(left, power);
   right = modify_inputs(right, power);
-
-  if(isdriver == false)
-  {
-    left_motors.spin(directionType::fwd, left * 12, voltageUnits::volt);
-    right_motors.spin(directionType::fwd, right * 12, voltageUnits::volt);
-  }else
-  {
-    left_motors.spin(directionType::fwd, left * 100.0, percentUnits::pct);
-    right_motors.spin(directionType::fwd, right * 100.0, percentUnits::pct);
-  }
+  left_motors.spin(directionType::fwd, left * 12, voltageUnits::volt);
+  right_motors.spin(directionType::fwd, right * 12, voltageUnits::volt);
 }
 
 /**
@@ -329,7 +321,8 @@ bool TankDrive::turn_to_heading(double heading_deg, double speed)
  */
 double TankDrive::modify_inputs(double input, int power)
 {
-  return (power % 2 == 0 ? (input < 0 ? -1 : 1) : 1) * pow(input, power);
+  if(power % 2 == 0 && input < 0) return -1 * pow(input, power);
+  return pow(input, power);
 }
 
 bool TankDrive::pure_pursuit(std::vector<PurePursuit::hermite_point> path, double radius, double speed, double res, directionType dir) {
